Redesign
Captured standards check code and put in new py file to preserve it for future development. Removed code related to standards check from main scripts. Refactored syntax, particularly print format statements.  Renamed mane script.
