+++ README.md
@@ -1,6 +1,9 @@
 # EnterpriseGDBInventory
-Inventory the content of an ArcSDE Enterprise Geodatabase.
-20180118: Created branch for a redesign. Redesign will create environment independence. 
+Inventory the content of an ArcSDE Enterprise Geodatabase. Inventory feature classes, domains, and fields within
+feature classes as well as details on each of these three elements. Write content to csv files.
+
+20180118: Created branch for a redesign. Redesign will create environment independence, and remove functionality
+for checking data against standards as this varies greatly with the user. 
 New script will only process one sde file rather than any sde files in the directory of choice.
 Script will no longer look for required fields, as they were employer specific.
