+++ TESTING_Results/20180201_iMAPDev_as_SDE.sde__FeatureClassFIELDSInventory.csv
@@ -1,0 +1,38 @@
+FIELD_ID,FC_ID,FLD_ALIAS,FLD_NAME,FLD_TYPE,FLD_DEF_VAL,FLD_DOMAIN,FLD_ISNULLABLE,FLD_LENGTH,FLD_PRECISION,FLD_SCALE,FLD_REQUIRED
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP.OBJECTID,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP,GIS Object ID,OBJECTID,OID,None,,False,4,10,0,True
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP.Acres,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP,Acres,Acres,Double,None,,True,8,38,8,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP.COUNTY,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP,County,COUNTY,String,None,,True,50,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP.AREA_NAME,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP,Area Name,AREA_NAME,String,None,,True,50,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP.UPDATED,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP,Updated,UPDATED,Date,None,,True,8,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP.JURSCODE,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP,Jurisdiction Code,JURSCODE,String,None,,True,4,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP.Shape,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP,GIS Shape,Shape,Geometry,None,,True,8,0,0,True
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP.Shape.STArea(),iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP,Shape.STArea(),Shape.STArea(),Double,None,,False,0,0,0,True
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP.Shape.STLength(),iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP,Shape.STLength(),Shape.STLength(),Double,None,,False,0,0,0,True
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.OBJECTID,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,GIS Object ID,OBJECTID,OID,None,,False,4,10,0,True
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.PROT,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,Type of Protection,PROT,String,None,,True,6,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.ACCTID,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,Account ID,ACCTID,String,None,,True,16,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.ACRES,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,Acres,ACRES,Double,None,,True,8,38,8,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.JURSCODE,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,Jurisdiction Code,JURSCODE,String,None,,True,4,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.COUNTY,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,County,COUNTY,String,None,,True,15,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.MDP_UPDATE,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,MDP Update,MDP_UPDATE,Date,None,,True,8,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.SOLD_DATE,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,Sold Date,SOLD_DATE,Date,None,,True,8,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.EASE_HOLDE,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,Easement Holder,EASE_HOLDE,String,None,,True,50,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.EASE_NUM_N,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,Easement Number,EASE_NUM_N,String,None,,True,50,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.COMMENT,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,Comment,COMMENT,String,None,,True,120,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.Shape,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,GIS Shape,Shape,Geometry,None,,True,8,0,0,True
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.Shape.STArea(),iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,Shape.STArea(),Shape.STArea(),Double,None,,False,0,0,0,True
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD.Shape.STLength(),iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,Shape.STLength(),Shape.STLength(),Double,None,,False,0,0,0,True
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.OBJECTID,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,GIS Object ID,OBJECTID,OID,None,,False,4,10,0,True
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.PROT,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,Type of Protection,PROT,String,None,,True,6,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.ACCTID,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,Account ID,ACCTID,String,None,,True,16,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.ACRES,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,Acres,ACRES,Double,None,,True,8,38,8,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.JURSCODE,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,Jurisdiction Code,JURSCODE,String,None,,True,4,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.COUNTY,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,County,COUNTY,String,None,,True,15,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.MDP_UPDATE,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,MDP Update,MDP_UPDATE,Date,None,,True,8,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.SOLD_DATE,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,Sold Date,SOLD_DATE,Date,None,,True,8,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.EASE_HOLDE,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,Easement Holder,EASE_HOLDE,String,None,,True,50,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.EASE_NUM_N,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,Easement Number,EASE_NUM_N,String,None,,True,50,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.COMMENT,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,Comment,COMMENT,String,None,,True,120,0,0,False
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.Shape,iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,GIS Shape,Shape,Geometry,None,,True,8,0,0,True
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.Shape.STArea(),iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,Shape.STArea(),Shape.STArea(),Double,None,,False,0,0,0,True
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP.Shape.STLength(),iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,Shape.STLength(),Shape.STLength(),Double,None,,False,0,0,0,True
