+++ TESTING_Results/20180201_iMAPDev_as_SDE.sde__FeatureClassInventory.csv
@@ -1,0 +1,4 @@
+FC_ID,ADM_ID,FC_FDNAME,FC_NAME,FC_DATATYPE,FC_SHAPETYPE,FC_SPATIALREFNAME,FC_DATEEXPORT
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PriorityPreservationAreas_MDP,iMAPDev_as_SDE.iMAPDev,_,SDE.AGRL_PriorityPreservationAreas_MDP,FeatureClass,Polygon,WGS_1984_Web_Mercator_Auxiliary_Sphere,02/01/2018
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP_OLD,iMAPDev_as_SDE.iMAPDev,_,SDE.AGRL_PermPreservedAgLands_MDP_OLD,FeatureClass,Polygon,WGS_1984_Web_Mercator_Auxiliary_Sphere,02/01/2018
+iMAPDev_as_SDE.iMAPDev._.SDE.AGRL_PermPreservedAgLands_MDP,iMAPDev_as_SDE.iMAPDev,_,SDE.AGRL_PermPreservedAgLands_MDP,FeatureClass,Polygon,WGS_1984_Web_Mercator_Auxiliary_Sphere,02/01/2018
