+++ TESTING_Results/20180201_iMAPDev_as_SDE.sde__GeodatabaseDomainsInventory.csv
@@ -1,0 +1,1 @@
+DOMAIN_ID,ENV_ID,DOM_NAME,DOM_OWNER,DOM_DESC,DOM_DOMAINTYPE,DOM_TYPE,DOM_CODEDVALKEYS,DOM_CODEDVALVALUES,DOM_RANGE,DOM_DATEEXPORT
